feat(dashboard): dashboard page router
